Resolve Bug
